requirements.txt
python-telegram-bot>=v20.3
peewee>=3.16.2
aiosqlite